+++ src/actions/index.js
@@ -1,17 +1,13 @@
-/**
- * Created by thomashourlier on 2/26/17.
- */
+import axios from 'axios'
 
-import axios from 'axios';
-
-export function fetchJedi() {
+export const fetchJedi = () => {
   return (dispatch) => {
     axios.get('http://localhost:3001/jedi')
       .then((res) => {
         dispatch({
           type: 'FETCH_FINISH',
           payload: res.data,
-        });
+        })
       })
   }
 }
