+++ src/components/App.js
@@ -1,0 +1,22 @@
+import React, { Component } from 'react'
+import logo from './logo.svg'
+import './App.css'
+import ListJedi from './ListJedi'
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <div className="App-header">
+          <img src={logo} className="App-logo" alt="logo" />
+          <h2>Welcome to React</h2>
+        </div>
+        <div className='App-row'>
+          <ListJedi/>
+        </div>
+      </div>
+    );
+  }
+}
+
+export default App
