+++ src/reducers/index.js
@@ -1,0 +1,6 @@
+import { combineReducers } from 'redux'
+import jedi from './jedi'
+
+export default combineReducers({
+  jedi
+})
