+++ src/reducers/jedi.js
@@ -1,0 +1,13 @@
+const jedi = (state = [], action) => {
+  switch (action.type) {
+    case 'FETCH_FINISH':
+      return [
+        ...action.payload,
+        ...state
+      ]
+    default:
+      return state
+  }
+}
+
+export default jedi
