+++ src/store.js
@@ -1,9 +1,8 @@
-/**
- * Created by thomashourlier on 2/26/17.
- */
+import { createStore, applyMiddleware } from 'redux'
+import thunk from 'redux-thunk'
+import reducers from './reducers'
 
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import reducer from './reducer';
-
-export default createStore(reducer, applyMiddleware(thunk));
+export default createStore(
+    reducers,
+    applyMiddleware(thunk)
+)
